Use the 'h' value from the slider, not from the computed color
In my project I'm encountering a bug where the hue slider immediately jumps back to the leftmost position when you move it anywhere else if the color that is currently being selected sits on the grayscale (#ffffff, #ececec, etc). I think this is because the slider computes the color value and determines that there is no hue and thus the hue slider should sit all of the way to the left. The problem is, if the user wants to pick a color in the blue, they should be able to slide the hue over and then use the main color picker area, but the hue slider immediately jumps back after they finish sliding it over.

My recommendation via this PR is to save the current hue slider position on a hue change and reuse that when updating the slider position as long as it is available. This should make the hue slider stay where the user drags it until the color picker component destructs and reconstructs with a new color value, at which point it can recalculate the hue from the color and move the slider to the far left again.