Use sliderH when determining the hue on updateColorPicker
A follow-up to https://github.com/zefoy/ngx-color-picker/pull/100. Saving the 'h' value from the onHueChange event fixed the cursor positioning on the hue slider, but now the saturation 'slider' flashes back to all red when you move the hue slider for a color on the gray scale. This seems to be fixable via the same method, you just have to reuse the saved sliderH value when calculating the hue as well as the slider position.